Fix text property not being counted properly
Previously, it is trying to find the length of content in [prop(attr1=value1, attr2=value2, ...)] content
But all lines with text="..." have the text as value, not as content, so the value length should be counted instead. No lines including a text="..." property also include content, so before this fix all text="..." lines were not counted in wordcount.